Add basic musicbrainz API metadata provider
MusicBrainz will be one of two currently planned providers of metadata.
When importing music, a metadata provider will be used to attempt to
match a given file with a track/album/artist and retrieve all of the
associated metadata. That includes the artist, album, and track
information along with any cover art.