Add MetadataProvider trait and QueryExpr enum
The MetadataProvider trait abstracts functionality that all metadata
provider clients should have. Metadata providers are used during music
import to match a given file to a artist, release, and track. The trait
will definitely change, as it will need to provide methods for
retrieving resources by the provider-specific identifiers, in addition
to performing search queries.

The trait has to return (boxed) futures for any asynchronous functions
since async traits aren't supported yet.

The QueryExpr enum is a very basic and very initial attempt to abstract
elements of search queries common to all metadata providers. The syntax
is pretty clunky, and the whole type may turn out to be necessary, but
it's a first step.